+++ runtime/v2/CHANGELOG.md
@@ -22,6 +22,8 @@ Each entry must include the Github issue reference in the following format:
 
 ## [Unreleased]
 
+* [#23607](https://github.com/cosmos/cosmos-sdk/pull/23607) Register runtime services properly.
+
 ## [v2.0.0-beta.1](https://github.com/cosmos/cosmos-sdk/releases/tag/runtime/v2.0.0-beta.1)
 
 Initial tag of `cosmossdk.io/runtime/v2`.